feat: manual gateway checkout
