+++ src/pages/checkout/[cartId].tsx
@@ -21,7 +21,7 @@ import { OrderCompleteState } from "../../components/checkout/types/order-pendin
 import { PresentCartState } from "@elasticpath/react-shopper-hooks";
 import { CheckoutForm as CheckoutFormType } from "../../components/checkout/form-schema/checkout-form-schema";
 import OrderComplete from "../../components/cart/OrderComplete";
-import StripeTypeCheckoutForm from "../../components/checkout/StripeTypeCheckoutForm";
+import ManualTypeCheckoutForm from "../../components/checkout/ManualTypeCheckoutForm";
 
 interface ICheckout {
   cart: ResourceIncluded<Cart, CartIncluded>;
@@ -82,7 +82,7 @@ export const Checkout: NextPage<ICheckout> = () => {
                 colStart={{ base: 1, md: 1 }}
               >
                 {showEpStripePaymentGateway ? (
-                  <StripeTypeCheckoutForm
+                  <ManualTypeCheckoutForm
                     checkout={stripeIntent}
                     showCompletedOrder={showCompletedOrder(presentCart)}
                   />
